+++ src/codegen/expression.rs
@@ -18,10 +18,6 @@ impl Transpilable for Expression {
             Expression::Boolean(value) => String::from(if *value { "true" } else { "false" }),
             Expression::Identifier(value) => format!("{}", *value),
             Expression::FunctionCall(f) => f.transpile(),
-<<<<<<< HEAD
-            Expression::BinaryOperator(_, _, _) => {
-                todo!("BinaryOperator codegen is not implemented yet")
-=======
             Expression::BinaryOperator(left_expr, right_expr, operator) => {
                 format!(
                     "{}{}{}",
@@ -29,7 +25,6 @@ impl Transpilable for Expression {
                     operator,
                     right_expr.transpile()
                 )
->>>>>>> f71f9ab ((lazily) codegen parsed expressions. v0.0.9)
             }
             Expression::UnaryOperator(operator, expression) => {
                 format!("{}{}", operator, expression.transpile())