+++ src/syntax/expression/mod.rs
@@ -13,3 +13,9 @@ mod unary;
 pub fn try_parse(tokens: &Vec<Token>, pos: usize) -> ParseResult<Expression, ()> {
     return equality::try_parse(tokens, pos);
 }
+
+
+#[cfg(test)]
+mod tests {
+    
+}