+++ CHANGELOG.md
@@ -19,6 +19,9 @@
 
 ## v0.0.10
 
+- [x] Parse function call parameters
+- [ ] Codegen function call parameters
+- [ ] Parse function declaration arguments
 - [ ] Begin work on semantic analysis
 - [ ] Symbol table
 - [ ] Typecheck bindings